blacklist is now staff command and dm set to rex/dash
