+++ cogs/owner.py
@@ -3,8 +3,7 @@ Misc commands that I want to run
 """ 
 import discord 
 from discord.ext import commands 
-from .utils import checks 
-from .utils.functions import extract_id
+from .utils import checks
 
 
 class Owner(commands.Cog):
@@ -113,134 +112,3 @@ class Owner(commands.Cog):
                         pass # self.bot.logger.warning('Failed: {}'.format(err))
             await ctx.send('Fixed the db')
         return
-
-    """
-    BLACKLIST
-    """
-    @commands.group(aliases=['blgu'], pass_context=True)
-    @commands.is_owner()
-    async def blacklistglobaluser(self, ctx):
-        """Add or remove a user to blacklist global list.
-
-        Parameters
-        ----------
-
-        Returns
-        -------
-        """
-        if ctx.invoked_subcommand is None:
-            users = await self.bot.postgres_controller.get_all_blacklist_users_global()
-            if isinstance(users, type(None)):
-                users = []
-            if len(users) > 0:
-                title = 'Users in global blacklist'
-                desc = '<@'
-                desc += '>, <@'.join(map(str, users))
-                desc += '>'
-            else:
-                desc = ''
-                title = 'No users in global blacklist'
-            embed = discord.Embed(
-                title=title,
-                description=desc
-            )
-            await ctx.send(embed=embed)
-
-    @blacklistglobaluser.command(name='add', pass_context=True)
-    async def _blgua(self, ctx: commands.Context, *, uids: str=None):
-        """Add user to global blacklist.
-
-        Parameters
-        ----------
-        uids: str
-            List of id, comma separated
-
-        Returns
-        -------
-        """
-        added_users = []
-        msg = uids.replace(' ', '')
-        if ',' in msg:
-            users = [extract_id(x, 'member') for x in msg.split(',')]
-        else:
-            users = [extract_id(msg, 'member')]
-        users = [x for x in users if x != '']
-
-        try:
-            for user in users:
-                success = await self.bot.postgres_controller.add_blacklist_user_global(user)
-                if success:
-                    added_users.append(user)
-            if added_users:
-                self.bot.blglobal += list(map(int, added_users))
-                title = 'Users added into global blacklist'
-                desc = '<@'
-                desc += '>, <@'.join(map(str, added_users))
-                desc += '>'
-                embed = discord.Embed(
-                    title=title,
-                    description=desc,
-                )
-            else:
-                self.bot.logger.info(f'Error adding users to global blacklist')
-                return
-            await ctx.send(embed=embed)
-        except Exception as e:
-            self.bot.logger.info(f'Error adding users to global blacklist {e}')
-
-    @blacklistglobaluser.command(name='remove', aliases=['rem', 'del', 'rm'])
-    async def _blgur(self, ctx: commands.Context, *, uids: str=None):
-        """Removes a user from the blacklist.
-
-        Parameters
-        ----------
-        uids: str
-            List of id, comma separated
-
-        Returns
-        -------
-        """
-        removed_users = []
-        user_notfound = []
-        msg = uids.replace(' ', '')
-        if ',' in msg:
-            users = [extract_id(x, 'member') for x in msg.split(',')]
-        else:
-            users = [extract_id(msg, 'member')]
-        print(users)
-
-        try:
-            for user in users:
-                success = False
-                try:
-                    success = await self.bot.postgres_controller.rem_blacklist_user_global(user)
-                    if success:
-                        self.bot.blglobal.remove(int(user))
-                        removed_users.append(user)
-                    else:
-                        user_notfound.append(user)
-                except:
-                    user_notfound.append(user)
-
-            fields = []
-            if removed_users:
-                fields.append(['PASS', ', '.join([f'<@{x}>' for x in removed_users])])
-            if user_notfound:
-                fields.append(['FAIL', ', '.join([f'<@{x}>' for x in user_notfound])])
-            title = 'Users blacklisting'
-            desc = ''
-            embeds = discord.Embed(
-                title=title,
-                description=desc
-            )
-            for field in fields:
-                embeds.add_field(
-                    name = field[0],
-                    value = field[1],
-                    inline = True
-                )
-            await ctx.send(embed=embeds)
-        except Exception as e:
-            self.bot.logger.warning(f'Issue removing users from ' +
-                                    f'global blacklist: {e}')
-
